Move IEditorOverrideService to common
